Added endpoint and password ingestion (#17)
* Added Pwned Passwords SHA-1/NTLM endpoint

Added Azure Function endpoint at range/append which will take a JSON POST payload with three values - sha1Hash, ntlmHash and prevalence. These are all validated using existing or added functions.

* Added searching existing Azure Blob Storage files for matching hashes

This could probably be optimised, but the functionality is there and works fine so far

* Azure Blob Storage now is updated

If a new hash is found, it is inserted in order, otherwise the value of the prevalence is updated and the line is updated before being re-written to the file

* No longer return boolean value if the entry was added or updated

* Improved tolerance of invalid JSON input

* Endpoint now takes JSON array of PwnedPasswordAppend

* Migrated code to use Azure Table Storage

Both range/{hashPrefix} and range/append now use Azure Table Storage for storage backend.

* Added Azure Function to update Blob Storage daily

This is very much still WIP - optimisations must be performed to only update the blob items which have changed

* Fixed compile errors as a result of merge

* Optimised Table Storage requests

There's still two key outstanding problems - firstly, that searching by Timestamp is *extremely* slow in Azure Table Storage, and secondly that Queries are limited to 1,000 items - these need to be segmented as to avoid capping out at 1,000 items for both reads and writes

* Contents of Blobs are served, removed BlobStorage UpdateHash function

* Now uses ExecuteQuerySegmented to deal with cases of >1000 items

Azure Table Storage's ExecuteQuery function is limited by 1,000 entries in a single response. To deal with this, ExecuteQuerySegmented should be used instead

* Optimised looking up modified hash prefixes for daily blob file update

Rather than using the Modified Timestamp, another Storage Table is now utilised - this will be written to if the hash prefix doesn't already exist in it, making it considerably faster than looking up modified timestamps greater than the last function execution time.

* Changes specified by PR review

Now returns entries that failed to be added, minor optimisations and moved stopwatches to use Stopwatch::StartNow()

* Fixed minor bug with nullable bool value producing incorrect output

* Adapted NTLM hash function to use new optimised methods

IsNTLMHash now uses IsHexStringOfLength(32) and added NTLM hash tests

* Added Cloudflare cache invalidation

Uses the API call to purge_cache with the list of hash prefixes. This is batched to circumvent API rate limits. Error tolerance could be improved.

* Added idempotency by hashing IP and data provided

This uses the CF-Connecting-IP header combined with the string of the data to generated a unique identifier which is stored in Table Storage for a quick lookup

* Added local memory cache for idempotency

* Local cache fix

* Migrated to Queue-driven model

Upon validation, the items are checked for duplication. Then, they are pushed to the queue. A trigger will dequeue the message and insert it into the database

* host.json update

* Removed test causing Null reference exception

* Fixed QueueTrigger not executing, using newer Azure.Data.Tables library

* Fixed nullable warnings

* Blob is written to every time the hash prefix is modified

* host.json syntax fix